Scheduler testing
