+++ .sqlx/query-008095e57274d4e68f1187cc825f161bfa8723666cfc83d76c6d86151be5528a.json
@@ -1,0 +1,14 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "\n            UPDATE games \n            SET \n                status = 'finished',\n                updated_at = NOW()\n            WHERE id = $1\n            ",
+  "describe": {
+    "columns": [],
+    "parameters": {
+      "Left": [
+        "Uuid"
+      ]
+    },
+    "nullable": []
+  },
+  "hash": "008095e57274d4e68f1187cc825f161bfa8723666cfc83d76c6d86151be5528a"
+}
