--- .sqlx/query-022c6a2f9136d66d53882a422c0a448511e0ff9a68bc4e9f67906a2ebfab4012.json
@@ -1,23 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n        INSERT INTO league_seasons (id, league_id, name, start_date, end_date, evaluation_timezone, auto_evaluation_enabled, game_duration_minutes, created_at, updated_at)\n        VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)\n        ",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "Uuid",
-        "Uuid",
-        "Varchar",
-        "Timestamptz",
-        "Timestamptz",
-        "Varchar",
-        "Bool",
-        "Float8",
-        "Timestamptz",
-        "Timestamptz"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "022c6a2f9136d66d53882a422c0a448511e0ff9a68bc4e9f67906a2ebfab4012"
-}
