+++ .sqlx/query-2fb1f2de8b8127f1940ab7960daf9ff7ddd6d9ba42a376a1cb1775eff85fd373.json
@@ -1,0 +1,14 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "\n            UPDATE games \n            SET \n                status = 'in_progress',\n                updated_at = NOW()\n            WHERE id = $1\n            ",
+  "describe": {
+    "columns": [],
+    "parameters": {
+      "Left": [
+        "Uuid"
+      ]
+    },
+    "nullable": []
+  },
+  "hash": "2fb1f2de8b8127f1940ab7960daf9ff7ddd6d9ba42a376a1cb1775eff85fd373"
+}
