--- .sqlx/query-98d7401e03349fb2c9c5f924cedf412bd3b4c8093b95be94f7ad502d4dfd8987.json
@@ -1,14 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n            UPDATE games \n            SET \n                status = 'finished',\n                game_end_time = NOW(),\n                updated_at = NOW()\n            WHERE id = $1\n            ",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "Uuid"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "98d7401e03349fb2c9c5f924cedf412bd3b4c8093b95be94f7ad502d4dfd8987"
-}
